% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSI.R
\docType{data}
\name{HSI}
\alias{HSI}
\title{Hang Seng Index (Daily, xts)}
\format{
An \code{xts} object with 7214 daily observations and a single
column \code{^HSI} containing adjusted close prices in index points.
The time index spans from 1986-12-31 to 2015-12-31.
}
\source{
Yahoo Finance, downloaded on 2016-01-03 via \code{qrmtools::get_data()}.
}
\usage{
data("HSI")
}
\description{
Daily adjusted close prices of the Hang Seng stock index (ticker
symbol \code{^HSI}), from its first date of availability on Yahoo
Finance to 2015-12-31.
}
\details{
Originally distributed as \code{HSI} in the \strong{qrmdata} package
(Hofert, Hornik, & McNeil), ported into \pkg{smqf} so that the book's
examples remain reproducible without an extra dependency. The data is
redistributed here under the same GPL (\eqn{\ge 2}) license as
\strong{qrmdata}.
}
\examples{
data("HSI")
class(HSI)        # "xts" "zoo"
dim(HSI)
head(HSI)

}
\references{
Hofert, M., Hornik, K., & McNeil, A. J. \emph{qrmdata: Data Sets for
Quantitative Risk Management Practice},
\url{https://CRAN.R-project.org/package=qrmdata}.
}
\keyword{datasets}
\keyword{finance}
\keyword{time-series}
