% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smqf-package.R
\docType{package}
\name{smqf}
\alias{smqf-package}
\alias{smqf}
\title{smqf: Statistical Methods in Quantitative Finance}
\description{
`smqf` is an R package providing datasets and functions used in the
book "Statistical Methods in Quantitative Finance" Ardia (2025).
}
\section{Acknowledgments}{

The market datasets `SP500`, `DJ`, `DJ_const`, `FTSE`, `FTSE_const`,
`EURSTOXX`, `EURSTX_const`, `DAX`, `CAC`, `NIKKEI`, `SMI`, `HSI`,
`GOLD`, and `VIX` are redistributed from the **qrmdata** package by
Marius Hofert, Kurt Hornik, and Alexander J. McNeil under the same
GPL (\eqn{\ge 2}) license. The original data was collected by the qrmdata authors from
Yahoo Finance and FRED (via Quandl) on 2016-01-03. See
<https://CRAN.R-project.org/package=qrmdata> for the upstream package
and the individual dataset help pages for details.
}

\references{
David Ardia (2025). *Statistical Methods in Quantitative Finance*. CRC Press.

Hofert, M., Hornik, K., & McNeil, A. J. *qrmdata: Data Sets for
Quantitative Risk Management Practice*.
<https://CRAN.R-project.org/package=qrmdata>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ardiad/smqf}
  \item Report bugs at \url{https://github.com/ardiad/smqf/issues}
}

}
\author{
David Ardia
}
\keyword{internal}
