% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smqf-package.R
\docType{package}
\name{smqf}
\alias{smqf-package}
\alias{smqf}
\title{smqf: Statistical Methods in Quantitative Finance}
\description{
`smqf` is an R package providing datasets and functions used in the
book "Statistical Methods in Quantitative Finance" Ardia (2025).
}
\references{
\insertRef{Ardia2025}{smqf} \cr
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ArdiaD/smqf-package}
  \item Report bugs at \url{https://github.com/ArdiaD/smqf-package/issues}
}

}
\author{
David Ardia
}
\keyword{internal}
